+++ src/data/reducers/connectionStatus.js
@@ -1,0 +1,10 @@
+import { PING_RESPONSE } from '../actions/pingStudio';
+
+export default function connectionStatus(state = null, action) {
+  switch (action.type) {
+    case PING_RESPONSE:
+      return action.status;
+    default:
+      return state;
+  }
+}
