+++ src/data/reducers/index.js
@@ -1,0 +1,9 @@
+import { combineReducers } from 'redux';
+
+import connectionStatus from './connectionStatus';
+
+const rootReducer = combineReducers({
+  connectionStatus,
+});
+
+export default rootReducer;
